Add OVERLAPPED_DISPATCH feature
Currently a program needs to finish on all cores before the next program
starts. This means that cores that finish early can't get started on the next
program sooner, and there's always a roundtrip to the dispatch core before a
core can start the next program.

Add an OVERLAPPED_DISPATCH define to allow a kernel to start before the
previous program finishes. Initially, the program may not modify L1 or DRAM
data that the previous program is still using, and should not expect that
semaphores are initialized on other cores. The kernel can call the
global_program_barrier() at any point to wait for the previous program to
complete; after that point it may do all operations like normal.